refactor(file-structure): standardize file naming convention
- updated filenames to use kebab-case for consistency
- moved request schemas to a dedicated `schemas` directory
- added new dependencies: `@hono/swagger-ui` and `@hono/zod-openapi`
- updated imports to reflect new file paths and naming convention
- added swagger UI integration at route `/ui`
