+++ __tests__/e2e-tests/health-check.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, spyOn, test } from "bun:test";
 import app from "../../src/app.ts";
 import { db } from "../../src/database";
 import { SENSITIVE_SERVER_DATA_AUTHORIZATION_TOKEN } from "../../src/env.ts";
-import { ApiError } from "../../src/utils/ApiError.ts";
+import { ApiError } from "../../src/utils/api-error.ts";
 
 const urlPrefix = "/api/v1";
