+++ src/controllers/v1/health-check.controller.ts
@@ -2,8 +2,8 @@ import { sql } from "drizzle-orm";
 import type { Context } from "hono";
 
 import { db } from "../../database";
-import { ApiError } from "../../utils/ApiError.ts";
-import { ApiResponse } from "../../utils/ApiResponse.ts";
+import { ApiError } from "../../utils/api-error.ts";
+import { ApiResponse } from "../../utils/api-response.ts";
 import logger from "../../utils/logger.ts";
 
 export const serverHealthCheckController = (c: Context) => {
